+++ src/imio/directory/core/subscribers.py
@@ -1,15 +1,19 @@
 # -*- coding: utf-8 -*-
 
+from imio.directory.core.contents import IEntity
 from imio.directory.core.rest.odwb_endpoint import OdwbEndpointGet
+from imio.directory.core.utils import ENDPOINT_CACHE_KEY
 from imio.directory.core.utils import get_entity_for_contact
 from imio.directory.core.utils import get_entity_uid_for_contact
 from imio.smartweb.common.faceted.utils import configure_faceted
 from imio.smartweb.common.interfaces import IAddress
 from imio.smartweb.common.utils import geocode_object
+from imio.smartweb.common.utils import get_parent_providing
 from imio.smartweb.common.utils import remove_cropping
 from plone import api
 from plone.api.content import get_state
 from Products.DCWorkflow.interfaces import IAfterTransitionEvent
+from zope.annotation.interfaces import IAnnotations
 from zope.component import getMultiAdapter
 from zope.globalrequest import getRequest
 from zope.lifecycleevent import ObjectRemovedEvent
@@ -47,6 +51,8 @@ def added_contact(obj, event):
     if not obj.is_geolocated:
         # geocode only if the user has not already changed geolocation
         geocode_object(obj)
+    # @search endpoint invaldation
+    invalidate_endpoint_search_cache(obj)
 
 
 def modified_contact(obj, event):
@@ -70,6 +76,8 @@ def modified_contact(obj, event):
         request = getRequest()
         endpoint = OdwbEndpointGet(obj, request)
         endpoint.reply()
+    # @search endpoint invaldation
+    invalidate_endpoint_search_cache(obj)
 
 
 def modified_entity(obj, event):
@@ -89,6 +97,8 @@ def moved_contact(obj, event):
         request = getRequest()
         endpoint = OdwbEndpointGet(obj, request)
         endpoint.reply()
+    # @search endpoint invaldation
+    invalidate_endpoint_search_cache(obj)
 
 
 def removed_entity(obj, event):
@@ -175,3 +185,14 @@ def set_uid_of_referrer_entities(obj, container_entity):
         obj.selected_entities.append(rel.from_object.UID())
         obj._p_changed = 1
     obj.reindexObject(idxs=["selected_entities"])
+
+
+def invalidate_endpoint_search_cache(obj):
+    site = api.portal.get()
+    ann = IAnnotations(site)
+    entity = get_parent_providing(obj, IEntity)
+    # if we're curently removing an entity
+    # there is no more this entity here
+    if entity:
+        ann_full_key = f"{ENDPOINT_CACHE_KEY}{entity.UID()}"
+        ann[ann_full_key] = int(ann.get(ann_full_key, 0)) + 1