+++ src/imio/directory/core/utils.py
@@ -8,6 +8,9 @@ import base64
 import vobject
 
 
+ENDPOINT_CACHE_KEY = "imio.directory.search_cache_generation"
+
+
 def get_entity_for_contact(contact):
     obj = contact
     while not IEntity.providedBy(obj):